separated out component for adding activity to Routine
